% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBNI-package.R
\docType{package}
\name{BBNI-package}
\alias{BBNI}
\alias{BBNI-package}
\title{BBNI: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anson-li8.github.io/BBNI}
  \item \url{https://anson-li8.github.io/BBNI/}
  \item Report bugs at \url{https://github.com/anson-li8/BBNI/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
