% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{GenerateNetwork}
\alias{GenerateNetwork}
\title{Generate Initial Network Topology}
\usage{
GenerateNetwork(num.node)
}
\arguments{
\item{num.node}{An integer representing the total number of genes/nodes in the network.}
}
\value{
A square transition function matrix combining both the initial DAG topology and the randomly assigned Boolean logic functions (represented by integer codes 1-12).
}
\description{
Randomly generates an initial, legal directed acyclic graph (DAG) topology
$T$ and assigns a corresponding Boolean transition function $F$ to each node.
This ensures that the maximum in-degree for any node is bounded by 2, and
utilizes internal checks to guarantee the resulting structure contains no
directed cyclic loops.
}
\examples{
# Generate a true network topology and Boolean rules for 5 nodes
set.seed(123)
true_network <- GenerateNetwork(num.node = 5)
print(true_network)

}
